backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# Config / env
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Background scheduler (breach timer)
apscheduler==3.10.4

# Google Calendar (Stage 3 — install now, use later)
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0

# Testing (Stage 4)
pytest==8.3.3
httpx==0.27.2
